requirements.txt
# Core
annotated-types>=0.7.0
pytz>=2025.1
anyio>=4.10.0
beautifulsoup4>=4.13.5
blinker>=1.9.0
certifi>=2025.8.3
cffi>=2.0.0
charset-normalizer>=3.4.3
click>=8.2.1
cryptography>=46.0.1
deepl>=1.22.0
deprecation>=2.1.0
distro>=1.9.0
feedparser>=6.0.12
# PULSE-WS: Flask dependencies restored - using Flask only
h11>=0.16.0
h2>=4.3.0
hpack>=4.1.0
httpcore>=1.0.9
httpx>=0.28.1
hyperframe>=6.1.0
idna>=3.10
iniconfig>=2.1.0
itsdangerous>=2.2.0
Jinja2>=3.1.6
jiter>=0.11.0
MarkupSafe>=3.0.2
openai>=1.108.0
packaging>=25.0
pluggy>=1.6.0
postgrest>=2.19.0
pycparser>=2.23
pydantic>=2.11.9,<3.0
pydantic_core>=2.33.2
Pygments>=2.19.2
PyJWT>=2.10.1
python-dateutil>=2.9.0.post0
python-dotenv>=1.1.1
PyYAML>=6.0.2
realtime>=2.19.0
requests>=2.32.5
sgmllib3k>=1.0.0
six>=1.17.0
sniffio>=1.3.1
soupsieve>=2.8
storage3>=2.19.0
StrEnum>=0.4.15
supabase>=2.19.0
supabase-auth>=2.19.0
supabase-functions>=2.19.0
tqdm>=4.67.1
typing-inspection>=0.4.1
typing_extensions>=4.15.0
urllib3>=2.5.0
websockets>=15.0.1
# PULSE-WS: Werkzeug restored - using Flask only

# Flask dependencies
Flask>=3.0.0,<4.0
Werkzeug>=3.0.0,<4.0

# PULSE-WS: FastAPI dependencies removed - using Flask only

# Reactor Core dependencies
aiofiles>=24.1.0
psutil>=6.1.0

# Dev tools
black>=24.8.0
flake8>=7.1.1
ruff>=0.8.0  # Быстрый линтер
safety>=3.6.0  # Security vulnerability scanner
bandit>=1.7.0  # Security linter

# Testing
pytest>=8.4.2
pytest-asyncio>=0.23.8   # 🔑 добавлено для async-тестов
pytest-cov>=6.0.0        # Покрытие кода тестами
psutil>=6.1.0            # Управление процессами и портами

# Telegram bot
aiogram>=3.22.0

# Advanced Parser dependencies
aiohttp>=3.12.15           # Асинхронные HTTP запросы
news-please>=1.6.16        # Машинное обучение для извлечения контента
trafilatura>=2.0.0         # Надежное извлечение текста
autoscraper>=1.1.14        # Обученные шаблоны извлечения
tenacity>=9.1.0            # Retry логика
scikit-learn>=1.3.2        # ML модели для самообучения
pandas>=2.1.4              # Обработка данных
numpy>=1.24.3              # Численные вычисления

# ORM (опционально)
SQLAlchemy>=2.0.35